Reduce memory consumption by only storing the required query texts
Previously we would intermittently store all query texts we receive in
memory, and only free them once the snapshot transformation and sending
was done. During transformation we would often see significant compaction
due to the difference between a pgss queryid and the pg_query fingerprint.

This capitalizes on that difference, and groups the text by the pg_query
fingerprint earlier, to avoid keeping duplicates in memory.